+++ src/serialization/__serialize_24.cpp
@@ -28,9 +28,9 @@ namespace fertilized {
     if (always_register ||
         serialization_library_version >= 0) {
       ar.template register_type<ClassicTraining<
-              uint8_t,
-              int16_t,
-              int16_t,
+              int,
+              int,
+              uint,
               std::vector<float>,
               std::vector<float>
 	  >>();
@@ -46,23 +46,23 @@ namespace fertilized {
       const unsigned int &serialization_library_version);
 
     TemplateFuncExport DllExport std::string serialize(const ClassicTraining<
-              uint8_t,
-              int16_t,
-              int16_t,
+              int,
+              int,
+              uint,
               std::vector<float>,
               std::vector<float>
 	  > *, const bool &);
     TemplateFuncExport DllExport ClassicTraining<
-              uint8_t,
-              int16_t,
-              int16_t,
+              int,
+              int,
+              uint,
               std::vector<float>,
               std::vector<float>
 	  >* deserialize(std::stringstream &);
     TemplateFuncExport DllExport void deserialize(std::stringstream &, ClassicTraining<
-              uint8_t,
-              int16_t,
-              int16_t,
+              int,
+              int,
+              uint,
               std::vector<float>,
               std::vector<float>
 	  >*);
@@ -75,9 +75,9 @@ namespace serialization {
 
 template <>
 struct version<ClassicTraining<
-              uint8_t,
-              int16_t,
-              int16_t,
+              int,
+              int,
+              uint,
               std::vector<float>,
               std::vector<float>
 	  >> {
