+++ src/serialization/__serialize_33.cpp
@@ -28,8 +28,8 @@ namespace fertilized {
     if (always_register ||
         serialization_library_version >= 0) {
       ar.template register_type<ClassificationLeafManager<
-              uint8_t,
-              int16_t
+              float,
+              uint
 	  >>();
     }
   };
@@ -43,16 +43,16 @@ namespace fertilized {
       const unsigned int &serialization_library_version);
 
     TemplateFuncExport DllExport std::string serialize(const ClassificationLeafManager<
-              uint8_t,
-              int16_t
+              float,
+              uint
 	  > *, const bool &);
     TemplateFuncExport DllExport ClassificationLeafManager<
-              uint8_t,
-              int16_t
+              float,
+              uint
 	  >* deserialize(std::stringstream &);
     TemplateFuncExport DllExport void deserialize(std::stringstream &, ClassificationLeafManager<
-              uint8_t,
-              int16_t
+              float,
+              uint
 	  >*);
 }  // namespace fertilized
 
@@ -63,8 +63,8 @@ namespace serialization {
 
 template <>
 struct version<ClassificationLeafManager<
-              uint8_t,
-              int16_t
+              float,
+              uint
 	  >> {
     typedef mpl::int_<FERTILIZED_VERSION_COUNT> type;
     typedef mpl::integral_c_tag tag;
