+++ src/serialization/__serialize_42.cpp
@@ -29,8 +29,8 @@ namespace fertilized {
         serialization_library_version >= 0) {
       ar.template register_type<ClassificationThresholdOptimizer<
               uint8_t,
-              int16_t,
-              int16_t
+              float,
+              uint
 	  >>();
     }
   };
@@ -45,18 +45,18 @@ namespace fertilized {
 
     TemplateFuncExport DllExport std::string serialize(const ClassificationThresholdOptimizer<
               uint8_t,
-              int16_t,
-              int16_t
+              float,
+              uint
 	  > *, const bool &);
     TemplateFuncExport DllExport ClassificationThresholdOptimizer<
               uint8_t,
-              int16_t,
-              int16_t
+              float,
+              uint
 	  >* deserialize(std::stringstream &);
     TemplateFuncExport DllExport void deserialize(std::stringstream &, ClassificationThresholdOptimizer<
               uint8_t,
-              int16_t,
-              int16_t
+              float,
+              uint
 	  >*);
 }  // namespace fertilized
 
@@ -68,8 +68,8 @@ namespace serialization {
 template <>
 struct version<ClassificationThresholdOptimizer<
               uint8_t,
-              int16_t,
-              int16_t
+              float,
+              uint
 	  >> {
     typedef mpl::int_<FERTILIZED_VERSION_COUNT> type;
     typedef mpl::integral_c_tag tag;
