+++ src/serialization/__serialize_61.cpp
@@ -28,9 +28,9 @@ namespace fertilized {
     if (always_register ||
         serialization_library_version >= 0) {
       ar.template register_type<DirectPatchDifferenceSurfCalculator<
-              int,
-              int,
-              uint
+              float,
+              float,
+              int16_t
 	  >>();
     }
   };
@@ -44,19 +44,19 @@ namespace fertilized {
       const unsigned int &serialization_library_version);
 
     TemplateFuncExport DllExport std::string serialize(const DirectPatchDifferenceSurfCalculator<
-              int,
-              int,
-              uint
+              float,
+              float,
+              int16_t
 	  > *, const bool &);
     TemplateFuncExport DllExport DirectPatchDifferenceSurfCalculator<
-              int,
-              int,
-              uint
+              float,
+              float,
+              int16_t
 	  >* deserialize(std::stringstream &);
     TemplateFuncExport DllExport void deserialize(std::stringstream &, DirectPatchDifferenceSurfCalculator<
-              int,
-              int,
-              uint
+              float,
+              float,
+              int16_t
 	  >*);
 }  // namespace fertilized
 
@@ -67,9 +67,9 @@ namespace serialization {
 
 template <>
 struct version<DirectPatchDifferenceSurfCalculator<
-              int,
-              int,
-              uint
+              float,
+              float,
+              int16_t
 	  >> {
     typedef mpl::int_<FERTILIZED_VERSION_COUNT> type;
     typedef mpl::integral_c_tag tag;
