+++ src/serialization/__serialize_75.cpp
@@ -29,8 +29,8 @@ namespace fertilized {
         serialization_library_version >= 0) {
       ar.template register_type<EqualDistBagging<
               uint8_t,
-              uint8_t,
-              uint,
+              int16_t,
+              int16_t,
               std::vector<float>,
               std::vector<float>
 	  >>();
@@ -47,22 +47,22 @@ namespace fertilized {
 
     TemplateFuncExport DllExport std::string serialize(const EqualDistBagging<
               uint8_t,
-              uint8_t,
-              uint,
+              int16_t,
+              int16_t,
               std::vector<float>,
               std::vector<float>
 	  > *, const bool &);
     TemplateFuncExport DllExport EqualDistBagging<
               uint8_t,
-              uint8_t,
-              uint,
+              int16_t,
+              int16_t,
               std::vector<float>,
               std::vector<float>
 	  >* deserialize(std::stringstream &);
     TemplateFuncExport DllExport void deserialize(std::stringstream &, EqualDistBagging<
               uint8_t,
-              uint8_t,
-              uint,
+              int16_t,
+              int16_t,
               std::vector<float>,
               std::vector<float>
 	  >*);
@@ -76,8 +76,8 @@ namespace serialization {
 template <>
 struct version<EqualDistBagging<
               uint8_t,
-              uint8_t,
-              uint,
+              int16_t,
+              int16_t,
               std::vector<float>,
               std::vector<float>
 	  >> {
